+++ src/launcher/pack.rs
@@ -4,6 +4,7 @@ use options::pack::{Options};
 use config::settings::Settings;
 use launcher::build::build_container;
 use launcher::wrap::Wrapper;
+use process_util::convert_status;
 
 
 pub fn pack_command(settings: &Settings, args: Vec<String>)
@@ -21,5 +22,6 @@ pub fn pack_command(settings: &Settings, args: Vec<String>)
     let mut cmd: Command = Wrapper::new(Some(&ver), &settings);
     cmd.userns();
     cmd.arg("_pack_image").args(&args);
-    cmd.run()
+    cmd.status().map(convert_status)
+    .map_err(|e| format!("Error running `vagga_wrapperr _pack_image`: {}", e))
 }