+++ src/options/mod.rs
@@ -1,2 +1,4 @@
 pub mod build_mode;
+pub mod compression_type;
 pub mod pack;
+pub mod push;